Get rid of TODOs
