+++ lib/membrane/core/pipeline/spec_controller.ex
@@ -53,7 +53,6 @@ defmodule Membrane.Core.Pipeline.SpecController do
          :ok <-
            state
            |> Parent.ChildrenModel.get_children()
-           # TODO handle_linking_finished has to be moved to Core.Element from Element
            |> Bunch.Enum.try_each(fn {_name, %{pid: pid}} ->
              pid |> Message.call(:linking_finished, [])
            end),
